first

+++ src/main/java/com/example/sber_test/repository/TransactionRepository.java
@@ -3,5 +3,5 @@ package com.example.sber_test.repository;
 import com.example.sber_test.domain.Transaction;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TransactionRepository extends CrudRepository<Transaction, Long> {
+public interface TransactionRepository extends CrudRepository<Transaction, Integer> {
 }
